test_task/requirements.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.1
djangorestframework==3.13.1
idna==3.3
install==1.3.5
kombu==5.2.3
prompt-toolkit==3.0.24
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
